requirements.txt
PyPDF2==3.0.1
beautifulsoup4==4.12.2
requests==2.31.0
python-docx==0.8.11
openpyxl==3.1.5
lxml==4.9.3
